Remove unnecesary code and Keep the 3 JS files
