+++ docs/index.html
@@ -1,1 +1,1 @@
-<!doctype html><html lang="en"><head><meta charset="UTF-8"/><meta http-equiv="X-UA-Compatible" content="IE=edge"/><meta name="viewport" content="width=device-width,initial-scale=1"/><title>To do list</title><script defer="defer" src="main.js"></script></head><body><main class="main"><div class="main__div"><h2 class="main__h2 main__elements">Demo</h2></div><div class="main__input-wrapper"><input class="main__input main__elements" placeholder="Add to your list..." name="input" id="input"/> <button class="main__button">Add</button></div><ul class="list main__ul main__elements"></ul><a class="main__anchor main__elements" href="#">Clear all completed</a></main><script src="index.js"></script></body></html>
+<!doctype html><html lang="en"><head><meta charset="UTF-8"/><meta http-equiv="X-UA-Compatible" content="IE=edge"/><meta name="viewport" content="width=device-width,initial-scale=1"/><title>To do list</title><script defer="defer" src="main.js"></script></head><body><main class="main"><header class="main__div"><h2 class="main__h2 main__elements">Demo</h2></header><form class="main__input-wrapper"><input class="main__input main__elements" placeholder="Add to your list..." name="input" id="input"/> <button class="main__button">Add</button></form><ul class="list main__ul main__elements"></ul><a class="main__anchor main__elements" href="#">Clear all completed</a></main><script src="index.js"></script></body></html>
